Fix reinforcement army ratio calculation

+++ AI/Nullkiller2/Engine/PriorityEvaluator.cpp
@@ -840,7 +840,7 @@ public:
 		// TODO: Mircea: See how we can get some kind of balance between MAINs in terms of army delivery
 		// See: GatherArmyBehavior::deliverArmyToHero
 		const uint64_t additionalArmyStrength = heroExchange.getReinforcementArmyStrength(evaluationContext.evaluator.aiNk);
-		const float additionalArmyRatio = additionalArmyStrength / heroExchange.hero->getArmyStrength();
+		const float additionalArmyRatio = static_cast<float>(additionalArmyStrength) / heroExchange.hero->getArmyStrength();
 
 		evaluationContext.addNonCriticalStrategicalValue(additionalArmyRatio);
 		evaluationContext.armyGrowth = additionalArmyStrength;
